feat: CRUD student and fix data type

+++ src/main/java/com/ra/base_spring_boot/repository/IndustryRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Set;
 
 @Repository
-public interface IndustryRepository extends JpaRepository<Industry, Integer> {
+public interface IndustryRepository extends JpaRepository<Industry, Long> {
     @Query("SELECT CASE WHEN COUNT(i) > 0 THEN true ELSE false END FROM Industry i WHERE i.id = :id")
     boolean existsByIdCustom(@Param("id") Long id);
     @Query("SELECT i FROM Industry i WHERE i.id IN :industryIds")
